+++ pigeons/native_reteno_plugin.dart
@@ -28,7 +28,7 @@ abstract class RetenoHostApi {
   List<NativeRecommendation> getRecommendations({
     required String recomVariantId,
     required List<String> productIds,
-    required String categoryId,
+    String? categoryId,
     List<NativeRecomFilter>? filters,
     List<String>? fields,
   });
@@ -198,12 +198,70 @@ class NativeRecommendation {
     this.price,
     this.description,
     this.imageUrl,
+    this.category,
+    this.categoryAncestor,
+    this.categoryLayout,
+    this.categoryParent,
+    this.dateCreatedAs,
+    this.dateCreatedEs,
+    this.dateModifiedAs,
+    this.itemGroup,
+    this.nameKeyword,
+    this.productIdAlt, // Alternative product_id field
+    this.tagsAllCategoryNames,
+    this.tagsBestseller,
+    this.tagsCashback,
+    this.tagsCategoryBestseller,
+    this.tagsCredit,
+    this.tagsDelivery,
+    this.tagsDescriptionPriceRange,
+    this.tagsDiscount,
+    this.tagsHasPurchases21Days,
+    this.tagsIsBestseller,
+    this.tagsIsBestsellerByCategories,
+    this.tagsItemGroupId,
+    this.tagsNumPurchases21Days,
+    this.tagsOldPrice,
+    this.tagsOldprice,
+    this.tagsPriceRange,
+    this.tagsRating,
+    this.tagsSale,
+    this.url,
   });
   final String productId;
   final String? name;
   final String? description;
   final String? imageUrl;
   final double? price;
+  final List<String>? category;
+  final List<String>? categoryAncestor;
+  final List<String>? categoryLayout;
+  final List<String>? categoryParent;
+  final String? dateCreatedAs;
+  final String? dateCreatedEs;
+  final String? dateModifiedAs;
+  final String? itemGroup;
+  final String? nameKeyword;
+  final String? productIdAlt; // Alternative product_id field
+  final String? tagsAllCategoryNames;
+  final String? tagsBestseller;
+  final String? tagsCashback;
+  final String? tagsCategoryBestseller;
+  final String? tagsCredit;
+  final String? tagsDelivery;
+  final String? tagsDescriptionPriceRange;
+  final String? tagsDiscount;
+  final String? tagsHasPurchases21Days;
+  final String? tagsIsBestseller;
+  final String? tagsIsBestsellerByCategories;
+  final String? tagsItemGroupId;
+  final String? tagsNumPurchases21Days;
+  final String? tagsOldPrice;
+  final String? tagsOldprice;
+  final String? tagsPriceRange;
+  final String? tagsRating;
+  final String? tagsSale;
+  final String? url;
 }
 
 enum NativeRecomEventType {